backend/requirements/base.txt
Django>=5.0,<6.0
djangorestframework>=3.14.0
drf-spectacular>=0.27.0
django-cors-headers>=4.3.0
psycopg2-binary>=2.9.9
python-dotenv>=1.0.1
Pillow>=10.2.0
django-environ>=0.11.2
djangorestframework-simplejwt>=5.3.1
uvicorn>=0.27.0
channels>=4.0.0
daphne>=4.0.0
channels-redis>=4.1.0
